Async all the things (well, almost)
Pushes async-style calling/execution right through from the web socket
layer right through up to the boundary with the DB.

Rationale being no need to block physical OS threads on DB
accesses. Things that block on DB access yield back to the tokio
scheduler which can then go do something else while they wait.

Caveats:

However, because of ownership model with Rocks, the DB itself still runs in a
separate thread with a mixture of crossbeam and tokio messages across
the boundary. This means a separate spawned thread for each Tx (though
they could in fact be pooled), so that's still a thing.

Additionally, having call-flow heavily asynchronous like this seems to
mess up the tracing/perfetto call-graph, making debugging harder. (This
is despite my best efforts to manually manage `tracing` spans to prevent
this.)

I'm not a huge fan of the call patterns for async, but it's probably easier
to come back from this than go forward doing the same thing *later* so here
we are.

(Ultimately it's going to come down performance vs ergonomics, so we
shall see in the long run if the numbers back it up.)