Rework verb calling flow:
Added a new VM return state, which returns (back to the task/scheduler) a continuation with the expected verb to call. The task then prepares the VM and dispatches it in the loop.

This will allow for more flexibility in dispatch, prepares the way for `fork` / `suspend`, and makes room for future support for verbs in multiple languages.

More importantly it will allow builtins to request execution of verbs; just as soon as I figure out a semantic for resuming them (likely through some sort of 'phase state' argument passed throug the bf call args. This is necessary for e.g. move() which has to call out to several MOO functions as part of its operation.