Implement bf_move, along with fixes we made along the way.
`move()` is possibly the most complicated builtin function, as it must dispatch out to a series of MOO-side verbs (:accept, :exitfunc, :enterfunc) to do some stuff along the way.

Uses the 'trampolining' facility for builtins to permit this call pattern.

Along the way I found bugs in the `caller` and `callers()` implementations.