Work towards getting `create` player working
A pile of fixes that fell out of turning over the rocks and finding
all the bugs that scattered out when I added a hook for the `create`
flow:

  * Fixes for regexp matching & subs.
  * Fixes for permissions handling in a pile of places.
  * Some missing/fixing builtins (`set_player_flags`, `value_bytes`,
    `object_bytes`, fixes to `connection_name`, fixes to `create`,
    `move`)

`create` now gets most of the way through, but dies in the last
instance when recording in the connection log. Almost there.