Yet another substantial reworking of permissions passing.
As a result of this pass, `create` of players now works. There were issues with property ownership and perms on lookup among other things.

Removed the PermissionsContext pieces which had both flags and the 'who', and was adding a bunch of complexity and leading to stale data.

Now the flags for 'who' are figured out inside the world state for each call.

It's quite likely there's still subtle bugs here, but it's closer to LambdaMOO's impl.