Fix mind bogglingly frustrating regression in [x..$] (`Op::Length`) type expressions inside exception handling blocks.
When I made exception handling's `PushLabel` functionality stop using the value stack, and use a separate 'handler' stack, I broke the calculation of value stack offsets used in the `Length` opcode.

This manifested itself as a runtime explosion while pulling in the offset for the value to use for calculating `$` in `[x..$]` type expressions, but only inside anything that used exception handling blocks, which use `PushLabel`.