Phase 1 of zero-copy around the data/storage-model.
This is the start of a broader attempt to move things from `binpack` to zero-copy value holders where it makes sense. The idea being that the on-disk and in-memory (and potentially network wire format) representation are the same, from back to front.

Conceptually, there is a shareable buffer type (SliceRef), which is reference counted and tied to its backing storage (for now `Arc<Vec<u8`...). Multiple SliceRefs from the same backing storage do not involve copies, and can be referred to by multiple parties without borrowing issues. When the last consumer drops its reference, it is freed.

For now this only applies to VerbDef, PropDef, the things that hold them, and sets of object ids (ObjSet). The layout is also subject to change. And because of limitations in the RocksDB Rust bindings API (see notes in SliceRef), it is not actually possible for the 0copying to be fully useful. There is still a copy (to a vec) made for each query.

The biggest remaining piece of (hard) work will be to transition Var itself to this model, where/if it makes sense. This is slightly tricky to implement. Also given the copy-on-write nature of MOO's type model (immutable values), it may make little sense from a performance reason anyways.