Major change to rework how the server is run/exposed:
Splits physical connection management and connection serving up so that they are in two separate OS processes: `daemon` for the latter and `host` for the former.

ZeroMQ is used to manage an RPC system between the two.

In theory it's possible for the `daemon` to go down and be brought up again without losing a connection managed by `host` (though this feature is not yet complete.)

As the # of crates has proliferated as a result, they've been moved into a `crates` subdirectory. A new README.md in there describes the layout.

(The old "monolithic" `server` binary still exists and runs, but will be deprecated.)