Rip out the bulk of anyhow usages inside vm/ and simplify error handling.
There was an incredible amount of pointless boilerplate here. In reality the VM always executes within a thread such that a panic is the most appropriate fail condition for situations that don't resolve into a MOO Error value.

So this commit changes the VM to either push an error to VM stack, or panic.