Implement `read()` builtin function. Allows for prompts.
https://github.com/rdaum/moor/issues/10

Like `suspend()` this commits the current transaction and then resumes in a new one once the input is back. So not really recommended if you care strongly about data consistency, and in general this kind of prompting breaks the "VR" feel.

Along with this I did a fairly major reworking of the scheduler loop, which should be far more efficient now. Was able to find a way to rework the channels etc. so that it works with a tokio select! invocation in a loop, so no need for a polling interval anymore.

(There's some oddities still that are preventing Session `fork` from
working that I need to get to the bottom of but they are non-critical;
and the experience in the console host isn't ideal.)