New, shinier, database.
Implements full persistence-backing for my bespoke 'tuplebox'; objects are flushed to backing-store (RocksDB to start) on commit.

Transactions see a snapshot-isolated view of the world through CoW persistent (in the FP sense) hashmaps (im::HashMap, a HAMT).

Replaces the "pure" RocksDB implementation for scenarios where the world size fits in RAM.

Lots of improvement here, but this is where we start.