Phase 1 of auth/auth work
All RPC requests to the daemon now have attached Client and Auth
PASETO tokens. The former is for validating that the RPC connection is
who it says it is. The latter is for validating that the user is who
they say they are.

It is now possible to get an auth token from logging in through
`/ws/auth` endpoint. Though there's nothing yet in place to use it.

The idea is that websocket connections (and HTTP generally as well)
will work this way:

1) auth and get a PASETO token for the player that is auth'd to.

2) subsequent calls to the system -- such as a websocket attach, or a
request to retrieve a property or verb, etc.) will attach the auth
token in a `X-Moor-Auth-Token` header on each request.

TODO:

  * validate the player inside the Scheduler/Task layer before
    launching a task (and return E_PERM etc.)

  * add a websocket connect that uses the authorization header to skip
    login, and remove the Basic-Auth (which can't work from a browser
    now anyways)

(https://github.com/rdaum/moor/issues/30)