Major rework of db layer + package upgrades
This is a big-bang merge of a bunch of major changes that happened on
a branch.

It rips out use of RocksDB entirely, replacing with custom page storage
layer and the use a proper write-ahead-log (via `okaywal` for now).

There's still some rough edges here, but this is much closer to the
vision I had in mind.