Optimize tuple access and make room for page-outs
Removes the pointless duplication of TupleRef & TupleId, and then
makes TupleRef hold a pointer to a backing "SlotPtr" entity which
points directly to the actual underlying tuple data in the buffer pool.

This acts both as an optimization -- avoiding a call through to the
"slotbox" each time to retrieve a tuple's value and giving a 20%
improvement in throughput in the db benchmark -- and a potential
future site for handling the swizzling of paged-out tuples. When
paging out, it should be possible to rewrite the SlotPtr at will, to
handle the movement of pages in and out of the buffer pool.