Add iterator support for the Adaptive Radix Tree
Also removes TupleKey, as it wasn't really going to be workable. Use ArrayKey or VectorKey and copy the data.

Support for actually using the ART as an index will require some refactoring of how values are typed & indexed, because the ART requires specific encoding for various types to make the partial encodings work; e.g. numeric values should be encoded Big Endian.

So the much-delayed work to add typing to the domain&codomain values is going to have to happen so that the indexing code can pick the right key-encoding method.