Refactor indexes to be "pluggable"
There's now an "Index" trait attached to relations presenting a
standard interface for implementing indexes. The two existing
implementations (CoW im::HashMap for canon base relations and HashMap
for working-sets) are moved into this.

This makes room for adding the Adaptive Radix Tree index, and maybe a
BTree one in the future.