fix(e2e): use `escargot` to build / look up all executed binaries
When I moved the e2e tests from `daemon` into `telnet-host`, I forgot to
update which binary gets build with `escargot` (required for cross-crate
builds), and which one we build "locally". This didn't explode because,
uh, shared build output and lucky build order and lack of hermetic
builds *handwaving intensifies*.

This resulted in: changes in code weren't picked up when re-running
tests locally after the code for `daemon` was modified.

Best to just build all binaries with `escargot` and stop thinking about
it.