Rework and fix how property definitions and values are handled
At some point values and properties got mixed up in how transitive
propagation and resolution was happening. PropDefs were being copied
down the inheritance hierarchy, but not values. Which is opposite of
how LambdaMOO has it.

This meant that bf_properties was returning the set of all properties
defined by parents instead of just the locally defined properties,
which was wrong.

This is a fairly major rework that rewrites things so that:

  PropDefs exist only on the root of the chain where the property is
    defined.
  Values propagate down the tree.
  Flags and owners are moved off PropDefs, and are separately defined,
    and again propagated down the tree
  Clearness remains defined by absence of a property value on a given
    point in the inheritance chain.

I could have broken this into more commits, but it kind of grew and
grew and grew...

STATUS: still some bugs in property resolution that causes JHCore to
choke on connect, but will get to the bottom of this later.