Initial implementation of moor database using `wiredtiger`
Adds an additional storage engine option, using WiredTiger

Splits up the database/worldstate storage model up so that more than
one backend is possible again.

Adds initial self-rolled wiredtiger bindings.

Implements the WorldState interface, and the connection DB on top of those bindings.

Some fundamental things unfinished:

Transaction model in WiredTiger is not really compatible with the moor
architecture as it is now. In particular, although it supports
snapshot isolation, it is a pessimistic model which will abort a
transaction immediately if a conflict could occur (instead of just at
commit time). To make this work I'll need to significantly alter how
transactions & verb executions are scheduled.

Secondly, wiredtiger has no support for monotonically incrementing
sequences, and so I'm emulating them using a set of atomic ints that
sync out to a table. It's not particularly robust.

Finally and most critically restoring a database doesn't actually work
yet, and I need to get to the bottom of this.

In any case, it's now available as a flavour of database to select at
startup (--db-flavour) for both the world state and connections
database.