Refactor worldstate backends to share most of the code
The WiredTiger and the "Relbox" backend were duplicating most of the
same code for mapping object values to/from tables.

This adds the addition of a of a common quasi-relational mapping trait
that is then implemented by both the wiredtiger and relbox
backends. Then the actual "worldstate" implementation is built overtop
of that.

A set of standard test functions can then be exercised against the implementation.