Fix tick slicing
The logic to aportion and then account for used ticks in the execution
loop was broken. This is what was breaking long running tasks like
e.g. "help intro."

At some point in some refactor I must have confused the idea of
`tick_slice` with `max_ticks` and the counting of the used ticks for
the task as a whole was being mixed up with the ticks for the current
pass through the execution loop. And so tasks were running forever not
being killed for hitting max count, but also treated further down as
if they had no ticks.