Fix "recycle" for when no ObjectParent tuple exists
a) Semantics for the wiredtiger mapping were kinda broken and were not
returning NotFound for remove-by-domain when the tuple did not
exist. So there was a difference in behaviour between WT and RB.

b) Missing propdefs and missing parent thus blew up in RB but not in
WT.

c) We were missing a test for recycle at the DB layer entirely.

There's more work to be done on proper full testing scenarios for
recycle and reparenting esp with inherited properties, but that's a
future-me problems.