Symbolization; symbols all the things
This is a rather big giant change to switch to using interned
strings (`Symbol`) where strings and string copies were used,
throughout.

So places where verb names or property names etc are referred to,
comparisons should be overall cheaper. But more importantly it adds
semantic clarity that the thing you're referring to is an identifier.

This does not (yet) present Symbol as a first class primitive type in
the MOO language, but it does change all property name and verb names
to be using them.

It also doesn't make it down to the DB layer... VerbDef, PropDef,
etc. are still using string values for storage and comparison, because
the problem of a transactional, concurrent-safe symbol table isn't
easy to solve right away.