Permit sending "rich" content via `notify`
Changes the RPC "Notify" event to carry a full MOO "Var" instead of a
string, along with a content-type.

The builtin function "notify" is now modified to take an optional
third argument, which is a content / mime-type. The intent here is to
allow cores & clients connected to them to do smarter things with
richer content.

For default (no argument), or for any content-type which starts with
"text/" the 2nd argument is expected to be a MOO string, same as
usual, and E_TYPE is raised for any other value types.

For any other content type, the value is allowed to be passed through
without modification and is up to the client to interpret it.

The existing telnet client just turns everything into strings, relying
on the default .to_string functionality on Var.