Add "TiedFlatBuffer" for managing flatbuffer encoded values
Takes the grunt work out of wrapping and passing around value objects that contain a flatbuffer.

Implements Clone and AsByteBuffer

Optimizes things a bit by only unpacking the root once, and then storing it (along with the Bytes data it came from) for later access.

Includes some basic support for data versioning that will be expanded.

This is in preparation for switching from bincode2 to flatbuffers throughout.