Rollback use of flatbuffers in values
This honestly ended up being a failed experiment and is getting in the way of a 1.0 release.

The original thought was to switch to a standardized off the shelf encoding format instead of the mix of bespoke binary encodings and `bincode` crate. But it's for many reasons simply not feasible at this time.

1. To switch now is simply too difficult

2. There are logistical problems with using flatbuffers throughout, especially around passing around nested values.

3. The API for flatbuffers is frankly difficult to work with and in the long run writing bespoke memory representations is simply easier in many respects.

This whole area continues to need attention. In particular I'm not super happy with the use of bincode.

I will also be reverting the use of flexbuffers for the `Var` values and switching back to much faster (but non-zero-copy, probably) values.