Adds "Fjall" as a DB backend for moor
Fjall is an LSM-based storage engine similar in some respects to
RocksDB, but written in pure Rust.

Depending on how this works out, this may become the default, and I
may retire the WiredTiger bindings, as it has become a source of
stress in compilation. And its isolation level is perhaps not high
enough.