Adds a new "model-checker" to do tx consistency check
https://github.com/rdaum/moor/issues/31

A program which runs a series of concurrent workloads of reading and
appending to lists on a set of random properties, in a number of
concurrent transactions. It then logs them into an "EDN" format file
that the `elle` tool from `jepsen` can read and check for
transactional isolation inconsistencies.

With this we are able to prove whether or not `moor` enforces strict
serializable isolation. At least in this (for now) limited use case.

Using this tool I was able to test the conflict-retry logic under
load, as well as find some bugs in it.

Testing shows that with the fjall-db backend and for this list-append
workload, moor behaves correctly with `serializable` isolation, but does
not conform to elle's definition of `strict-serializable`