Fix forever-blocking on results for retries on commit conflict
The TaskHandle for a task that is re-run due to commit-conflict would
sit blocking forever because the scheduler in fact has created a new
one.

This change propagates out a new task handle up to the daemon to let
it know that it needs to listen on that instead, and replace the the
task handle it was working with.

Found while doing concurrent tx model checking.