Major rework and simplification of the database layer
Replaces use of wiredtiger and custom relbox etc with a simplified
in-memory transactional front to non-transactional FjallDB storage.

This looks a lot like what the original moor database code looked like
two years ago,rips out a lot of code, simplifies a lot of things, and
most importantly makes things much much faster.

There is still a lot of tweaking to do here, but it passes the Jepsen
model check as serializable.

More on a new db