Adds a new "flyweight" type - a lightweight object
"Flyweights" are a bit of a fusion of anonymous objects and ColdMUD frobs.

They are composed of a sequence of "slots" which behave like
properties, a "delegate" to which verb and property (in case of slot
miss) delegate to, and a "contents" which is a sequence of values.

They are immutable and reference counted, and passed by value like MOO
lists and maps.

They will additionally (future) support a "sealing" function which
makes their contents opaque.

"this", and "caller" can be flyweights.

In the future object/room contents will permit flyweights. And command
matching will permit their return as well.