Fixes to multiline read() loops & pasting in client
Fix state management on the websocket read loop so that input is
spooled and not processed as commands until the currently submitted
task completes. RequestInput (`read()`) can pull from the spool.

This fixes the situation where a data race could break multiline input
or pastes from the client, causing them to be issued as new commands
because the new read() couldn't be issued before the host popped the
next line from the client.

Also adds explicit `paste` handling into the web client, improving the
UX