Allow return as expression, allows short-circuit conditionals
In Julia and some other languages, the following is allowed:

  true && return;
  player == this && return true;

etc

Which is a nice abbreviated idiom to avoid large chains of if blocks

This is possible because in those languages "return" is an expression
not a statement.

This is now the case for moor, too.

This also fixes old broken string literal parsing code which returned
E_INVARG instead of compiler error for invalid ints.