Major improvements to commit performance
There was a regression (due to some debugging code added earlier in
the week) that added a constant 10ms overhead on every single
commit. Fixed.

Also made the transaction commit reply process complete before the
sync to disk, since that's no business of the caller, and was adding a
constant overhead as well.

Extremely improved per-tx commit overhead, a factor of 100 improvement
in many cases.