A partial rewrite of the transaction layer
Uses im::HashMap behind the scenes to hold the "canonical" values
during reads, in hopes of reducing thrash on read retrievals.

Still holds a lock on final write completion, but this could
potentially be replaced with optimistic lock coupling or similar.

Eliminates background cache eviction, and eviction support generally,
with the assumption that we will not be working with databases larger
than main memory for many years.