Rework & fix transaction retry on commit-conflict
Formerly, when a transaction encountered conflict at commit time, the
whole task would be started from its "TaskStart" as if it was a brand
new command/verb/etc execution. This had two problems:

  * Was inaccurate/buggy because the actions performed by successful
  commit/suspend operations along the way weren't considered. So the
  whole task would restart as net-new even though it had partially
  succeeded.

  * Was slow as molasses because the whole task setup process was run
  again and again.

Now instead the state of the VM execution is snapshotted at task start
and at each successful commit, and we simply swap-in and restore the
snapshot so we restart back at the last successful commit.

This is dramatically faster, and more correcter.

https://github.com/rdaum/moor/issues/410