Add `active_tasks` to return running tasks
https://github.com/rdaum/moor/issues/406

Function: active_tasks()
Returns the list of active running (not suspended/queued) foreground tasks.
If the player is a wizard, it returns the list of all active tasks, otherwise it returns the list of
tasks only for the player themselves.
The information returned differs from queued_tasks and provides only:
{ task_id, player_id, task_start } where task_start is a description of how the task was started,
and varies depending on the type of task:
{ 'command, #handler, #player, "command" }
{ 'do_command, #handler, #player, "command" } - when $do_command has been invoked
{ 'verb, #player, #verb_location, 'verb, { args }, "argstr" }
{ 'eval, #player }
{ 'fork, #player, #perms, parent task, verb_location, verb_name, args }