Make the database store "programs" rather than binary blobs
Should speed up verb dispatch significantly.

In the past the thought here was that mooR's database could be agnostic
about program language / storage format, and so it would just store a
binary representation along with a "BinaryType" enum and let the runtime
decode on demand.

The problem is without sophisticated caching this means constantly
decoding the verb again and again on every dispatch. And as the database
already does its own caching, it is preferable to just move it up to
that level.

And we're not supporting multiple languages yet anyways.