Some cleanups on variable scoping (#417)
* Clean up variable decl / scopes

A bunch of this stuff got moved up into `common` when it strictly
speaking belongs in `compiler` alongside the abstract syntax tree.

This also clarifies some naming and cleans up the API a bit.

* Rework variable binding some more

Keep track of the type of variable binding.
Clarify some scoping

* Cleanup after clippy

* Fixup / cleanup scope management a bit

Do not emit scopes when the environment width for the local scope is
empty.