Perform "world state" actions on the scheduler in batch
Cleans up the Scheduler such that the execution of actions that are
not task related are done in a separate "world state executor", and
done potentially in batch. They are dispatched on their own thread.
This cleans up the main Scheduler source file, and makes room for
better management of these kinds of activities, and more of them.