Fix compile error results
LambdaMOO is capable of returning multiple compile errors from single
compilation, and set_verb_code and eval report all errors in a list.

mooR was incorrectly returning a string instead of a list for the
(single) compile error it was returning. So giving nonsensical results
in-core with messages like "48 errors."