Optimize for-seq/for-range loops
Instead of thrashing around on the value stack to hold the sequence
and offset, stores these in the scope stack and manipulates them
directly.

Shaves off some time unwrapping integer values and pushing/popping
from the value stack. Benchmark improvements were not dramatic, but
this sets the stage for future improvements.