Add `load_object` builtin
`load_object({objdef}, [dest_object | #-1, [constants_mapping]))`

Takes the result of a `dump_object` invocation (list of strings in
objdef format) and loads it into an object.

If the second argument is provided, the object is loaded at that
destination object ID. If it's #-1, a new object is created. Otherwise
the object # used in the dump itself is used.

If the third argument is provided it's treated as a mapping of
constant symbols to values/object ids, as in the `constants.moo` in an
objdef dump.

There is currently no attempt to manage conflicts if the inbound
properties or verbs already exist. Instead, an error is
thrown. Options for managing this are TBD.