Adjust for 1-indexing in E_RANGE errors in List
Unfortunately we don't know what the indexing mode is at the time we
execute here because that's handled by Var, but the vast majority of our
callers are MOO code that is 1-indexed, so for now we'll just report
the 1-indexed form.

At some point in the future we can look at passing the indexing mode
down, maybe.