Fix object matching in $mail_agent:parse_address_field
I switched this to using my new $code_utils:match_objid, but that only
matches exact strings while this code was expecting to find the object
id anywhere inside brackets in the string.

Switched to a manual regex which works for both normal numeric object
ids and uuid-object ids.

Fixes #504