Split raw line-mode TCP and telnet implementations
When a connection comes in, does a Do for Echo, and only uses the
telnet codec if the client responds.

This makes it so existing "raw" line mode TCP clients can continue to
work (modulo the first two Do bytes on the connection).

A pile of duplicated code between the two implementations right now,
but it works. Will have to refactor later.