Keep uuid objects out of $recycler
mooR doesn't need to use the $recycler for the reasons it was created in
LambdaMoo's core: performance and object ID conservation. Since there is
an issue with recreating UUID objects from the recycler (issue #513) we
decided to force uuid objects to avoid it completely regardless of the
build-option bi_create's setting.