Fix transaction bounary issues with `load_object`.
There was an issue where loading an object happens in a new
transaction but the existing transaction continues after, unable to
see (because of snapshot isolation) the results of the load.

This makes `load_object` return a TaskSuspend::Commit after it's done,
and then the task will resume in a new transaction, seeing the results
of the commit.

This also updates the semantics for TaskSuspend::Commit to take a
value, which `load_object` uses to return the results of the load.

Finally, this also changes `commit()` so that it takes an optional
value which is returned after the task resumes in the new transaction,
now that this facility exists.