Implement CurveZMQ encryption for daemon
Replaces the use of PASETO tokens for host/worker<->daemon
authentication with ZMQ's native "Curve" encyption and "ZAP"
authentication.

This makes it now so all traffic is end to end encrypted over the
network, including all RPC traffic.

It also makes the attaching of hosts and workers into the daemon into
a properly authenticated process which should make it impossible for
intruders (who do not have the enrollment key, which can be temporary)
to attach.

For ipc:// type sockets (on the same host), no auth or encryption is
applied.

For tcp:// type addresses, a shared enrollment token can be used
initially to register hosts with the daemon and exchange keys.

Hosts and workers no longer need to have key files shared with the
daemon, and the PEM files which the daemon only uses for PASETO token
generation for logins exist only on the daemon.

See docker-compose.yaml and accompanying documentation for how the
enrollment process works.