Queue all tasks
Cleans up / simplifies the task submission process to put all tasks
into the task q even at initial start.

While this will likely add a few nanoseconds of latency (measures
about 10ns on my machine) the advantage is a simplified and cleaner
code path and the ability to do more intelligent scheduling in the
future should we want to.