Initial version of "webhooks" API into the web-host
Can be used to serve arbitrary content over HTTP from the running MOO.

If the webhooks argument is true, and
There is a verb on `#0:invoke_http_handler`
Paths like `/webhook/xxxx` will call it and pass a series of
values, headers, etc in.
And then the results of that method invocation can be returned back to
the caller.