Adds much improved compile error reporting
The CompileError now provides extended diagnostics which can be used by
the caller to provide richer reporting.

In the objdef loader, this is used along with ariadne to produce a nice
report.

In set_verb_code and eval the default option is normal text renderable
with different levels of diagnostics and optional colour/ANSI graphics
lervels.