Begin work on cleaning up and deployment configs
This is an effort (re #545) to create a clearer setup for new or
existing users to have reference deployment configurations for mooR,
and consists of the following:

Providing clear multiple scenarios -- docker compose telnet only,
docker compose simple-web, web + ssl, debian packages, and potentially
a clustered (k8?) configuration in the future.

As such there's different directories for each of those, with
single-host but orchestrated deployments that use IPC sockets so don't
need host etc. enrollment.

Also modifies the root docker compose, process-compose etc to simplify
them by removing TCP sockets and enrollment tokens and the like from
the default development setup. (But adds a docker compose file to be
used specifically for testing that.)