Clarify how calling lambdas on properties is broken
I tried to make the compiler allow this:

some_obj.my_property = fn (test) return test; endfn;
some_obj.my_property(123);

But I ran into all sorts of parsing issues and it's not making it
work. Maybe someday we'll replace pest and when we do that we'll come
up with a better compiler behaviour.

In the interim, the book will tell you not to do this.

Fixes #572