DB: Properly check fully loaded before operations
We were not consistently applying the check for "fully loaded" before
hitting the "provider" looking for unfilled data.

This meant that we were needlessly hitting the disk to check for
tuples that might not be present in local cache, leading to pessimal
performance esp when there's a slow disk

15%ish performance improvement on my own local workload here, but
would potentially be much higher if the machine is using spinny spinny
old slow disks.