Fix null JSON values to behave like ToastStunt in parse_json
MOO doesn't have a "null" or "none" value. So we can't translate
e.g. { x: null } into anything meaningful.  ToastStunt crudely turns
these into "null" which is frankly about as good as we're going to get
unless we raise an error or come up with something clever. For the
same of compatibilty, I'll do the same.