Optimize a bunch of type dispatch logic
We were using .variant() in a lot of places which creates a wrapper
object around the underlying value. This wasn't an issue before sine Var
 already stored the Variant, but now Variant is purely synthetic.

However matching on .variant() creates a jump table in assembly which
isn't always the most efficient, esp for the common/hot path.

This commit changes things so that we try the as_xxx forms first for
various hot paths (like int int addition, etc.).