Add a verb suggestions palette & auto-say mode
Client can now query the server to ask what environmental verbs are
available, or use a default list.  If the verb palette setting is on,
a set of pills is available for users to tap and use.

Additionally, if "say mode" is on, by default every input defaults to
"say" unless the user hits backspace